chore: 기존 의미 없는 `tickerService` 에서 `CreateTIckerUseCase`로 서비스 레이어 이동
