+++ src/main/java/com/joojoo/api/ticker/application/TickerService.java
@@ -2,7 +2,4 @@ package com.joojoo.api.ticker.application;
 
 public interface TickerService {
     void addStockToRedis(String name, String ticker);
-
-    // 운영 DB에 있는 주식을 Redis에 저장
-    void loadTickersToCache(Long userId);
 }
